requirements.txt
Flask==0.10.1
Flask-Login==0.3.2
Flask-Migrate==1.6.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.12
GeoAlchemy2==0.2.6
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
Pillow==2.6.0
SQLAlchemy==1.0.10
SQLAlchemy-Utils==0.31.4
Shapely==1.5.13
WTForms==2.1
WTForms-Alchemy==0.13.1
WTForms-Components==0.9.8
WTForms-JSON==0.2.10
Werkzeug==0.11.2
alembic==0.8.3
aniso8601==1.1.0
argparse==1.2.1
blinker==1.4
decorator==4.0.6
fake-factory==0.5.3
gunicorn==19.3.0
infinity==1.3
intervals==0.3.1
itsdangerous==0.24
meld3==1.0.2
passlib==1.6.5
psycopg2==2.6
python-dateutil==2.4.2
python-editor==0.5
pytz==2015.7
six==1.10.0
supervisor==3.1.3
validators==0.9
wsgiref==0.1.2
